removed reconciler factory
On-behalf-of: @SAP a.shcherbatiuk@sap.com
