+++ cmd/gateway.go
@@ -11,7 +11,6 @@ import (
 	"github.com/openmfp/golang-commons/sentry"
 	"github.com/spf13/cobra"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
 	"github.com/openmfp/golang-commons/logger"
 
@@ -45,7 +44,7 @@ var gatewayCmd = &cobra.Command{
 			defer openmfpcontext.Recover(log)
 		}
 
-		ctrl.SetLogger(zap.New(zap.UseDevMode(true)))
+		ctrl.SetLogger(log.Logr())
 
 		gatewayInstance, err := manager.NewGateway(log, appCfg)
 		if err != nil {
